#170 Add second senzing_governor.py

+++ init-container.py
@@ -50,7 +50,7 @@ except Exception:
 # Metadata
 
 __all__ = []
-__version__ = "1.8.0"  # See https://www.python.org/dev/peps/pep-0396/
+__version__ = "2.0.0"  # See https://www.python.org/dev/peps/pep-0396/
 __date__ = '2019-07-16'
 __updated__ = '2022-07-01'
 
@@ -437,6 +437,7 @@ message_dictionary = {
     "184": "Initializing for MS SQL",
     "185": "Initializing for MySQL",
     "186": "Initializing for PostgreSQL",
+    "187": "{0} - Directory does not exist; no change.",
     "292": "Configuration change detected.  Old: {0} New: {1}",
     "293": "For information on warnings and errors, see https://github.com/Senzing/stream-loader#errors",
     "294": "Version: {0}  Updated: {1}",
@@ -1536,14 +1537,17 @@ def database_initialization_postgresql(config):
 
 def install_senzing_postgresql_governor_file(config, senzing_governor_path):
     if not os.path.exists(senzing_governor_path):
-        governor_url = config.get("governor_url")
-        logging.info(message_info(180, senzing_governor_path, governor_url))
-        try:
-            urllib.request.urlretrieve(
-                governor_url,
-                senzing_governor_path)
-        except urllib.error.URLError as err:
-            logging.warning(message_warning(301, governor_url, err))
+        if os.path.exists(os.path.dirname(senzing_governor_path)):
+            governor_url = config.get("governor_url")
+            logging.info(message_info(180, senzing_governor_path, governor_url))
+            try:
+                urllib.request.urlretrieve(
+                    governor_url,
+                    senzing_governor_path)
+            except urllib.error.URLError as err:
+                logging.warning(message_warning(301, governor_url, err))
+        else:
+            logging.info(message_info(187, senzing_governor_path))
     else:
         logging.info(message_info(181, senzing_governor_path))
